requirements.txt
mcp>=1.0.0
yt-dlp>=2024.1.1
openai-whisper>=20231117
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0
scipy>=1.11.0
scenedetect[opencv]>=0.6.0
opencv-python>=4.8.0
